chore(fe): move Icon file directory location
